COM-1695: move add/update/delete photo logic
